refactor: optimize compile speed (skip fetch deps)
Optimization takes use of the `skip-fetch-latest-git-deps` compilation flag.
Regardless of this flag, by default - if no deps are found in the
`/Users/$USER/.move` directory (a cache folder for Move packages)
the missing packages will be downloaded there.

But if the cache exists, then the `skip-fetch-latest-git-deps` flag can
be used to skip checking whether those cached Move packages ought to be
updated.

In this PR - for the first compilation step (checking that the original
tests are working fine), we disable that flag and ensure we get the
latest Move packages.
Then, for all subsequent compilations (mutant generation and running
tests on those mutants), the tool enables that flag since it doesn't
make much sense to check the latest deps repeatedly - checking it only
once when the tool is started - it should be good enough.