chore: add integration tests for spec test
Added the first integration tests for the `move-spec-test` that are executed on these projects:
- move-mutator/tests/move-assets/same_names
- move-mutator/tests/move-assets/poor_spec

The expectation for the integration test is that the generated
report by the move-spec-test will match the pre-generated
reports in those projects that end with the `.spec-exp` suffix.

Prefix for move-mutation-test has been update to `.mutation-exp` to
avoid collision in the projects that are used in both integration test
suites.

In a future change, we will add `spec-test` to the CI and also enable a
few more projects in the integration tests (e.g. the `simple` project).